code reorganization pre call options
